- minor folder organization, moved components to ui foldler
- updated imports paths

- updated todo
