+++ frontend/src/components/ui/Hero/Hero.jsx
