+++ frontend/src/components/ui/Hero/Hero.module.scss
