+++ frontend/src/components/ui/Rating/Rating.module.scss
