+++ frontend/src/components/ui/StatCard/StatCard.module.scss
