+++ frontend/src/pages/Home/Home.jsx
@@ -5,7 +5,7 @@ import styles from './Home.module.scss';
 
 import Spinner from '@components/common/Spinner/Spinner';
 import Button from '@components/common/Button/Button';
-import Hero from '@components/common/Hero/Hero';
+import Hero from '@components/ui/Hero/Hero';
 import SidePanel from '@components/common/SidePanel/SidePanel';
 import Icon from '@components/common/Icon/Icon';
